Fixes
for retro achievement progress
